change parseInt to parseFloat in inputStrValidation() function to preserve decimal places

+++ utilities/inputStrValidataion.js
@@ -14,13 +14,13 @@ export function inputStrValidation(cmdLineStr, currentNumArr = [], currentOperAr
             // how to undo the operators already pushed into currentOperArr? 
 
             currentOperArr.push(currentEl)
-        } else if (!operInputStarted && parseInt(currentEl)) {
+        } else if (!operInputStarted && parseFloat(currentEl)) {
             // if current char is a number and operInputStarted is false, push char to number array
-            currentNumArr.push(parseInt(currentEl))
-        } else if (operInputStarted && !parseInt(currentEl)) {
+            currentNumArr.push(parseFloat(currentEl))
+        } else if (operInputStarted && !parseFloat(currentEl)) {
             // if operInputStarted is true, more operators push into oper array 
             currentOperArr.push(currentEl)
-        } else if (operInputStarted && parseInt(currentEl)) {
+        } else if (operInputStarted && parseFloat(currentEl)) {
             // if operIuptuStarted is true and current char is a number. reset the currentNumArr and currentOperArr. remind user to check the cmd-line input and corrected errors. 
             // break current loop
             currentNumArr.splice(0)
